Add files via upload
Place this script in Applications > After Effects YYYY > Scripts. Run this in an empty Composition. It will ask you to select a folder on your computer with video files. Then it will go one-by-one, making a new layer in the composition containing that video, and automatically trim it to one second. It will sequentially stack them, stairstep-style, so you can immediately scrub and begin to modify the video inside the trim to get the best moment.