Add files via upload
Place this script in Applications > After Effects YYYY > Scripts

Prepare a txt (not RTF) file containing one line for each caption you want to fade in and out in your video.

When run the script will ask you to locate the text file on your computer, and then will create a next text layer called “Captions” (it may append an incremented index if any already exist), and make a fade in and fade out marker for each line in the txt file. It also adds a layer called “Globals” (or adds to any existing layer with the same name) by which you can tweak aspects of the fade in and fade out behavior. Then you can begin to manuallly adjust the markers to where they make sense in the composition. Should save a bunch of tedious production time.